requirements.txt
requests
python-dotenv
colorama
pyfiglet